requirements.txt
python-multipart==0.0.5
gputil==1.4.0
asyncio==3.4.3
async-timeout==3.0.1
pymongo==3.10.1
azure-cosmos==4.0.0
azure-storage-blob==12.8.1
azure-storage-queue==2.1.0
opencv-python==4.2.0.34
xxhash==1.4.3
aiohttp==3.7.4
psutil==5.7.2
aiofiles==0.5.0
redis-py-cluster==2.0.0
flask_cors==3.0.9
more_itertools==8.5.0
shapely==1.7.1
pyzmq==20.0.0
decorator==4.4.2
attrs==20.3.0
scipy==1.6.1
pytest==6.2.2
onnx==1.9.0
tf2onnx==1.8.5
pydantic==1.8.2
pyhumps==3.0.2
lru-dict==1.1.7
onnxruntime-gpu==1.10
tensorflow==2.5
fastapi==0.72.0
#uvicorn==0.14.0
uvicorn[standard]==0.17.0
torch==1.10.1
torchvision==0.11.2
#transformers>=4.0.0
transformers==4.16.1
SentencePiece==0.1.96
Pillow==9.0.0
aiofile==3.7.4